Arduino_GFX: Export writeAddrWindow()
For performance reasons (like in LVGL...) someone may want to directly
send pixel data to the screen, skipping all the boundary checks from
other methods of this library (as the checks were already performed
elsewhere, that would be double checking).

Make writeAddrWindow() usable from an external program, so that it
becomes possible to get a slight performance increase in some use cases.